Refact - Deep 모델의 view,serializer class 위치 변경 (#31)
