+++ backend/src/api/serializers.py
@@ -1,6 +1,5 @@
 from django.contrib.auth.models import User, Group
 from rest_framework import serializers
-from deep.models import Deep
 
 # User
 class UserSerializer(serializers.ModelSerializer):
@@ -13,13 +12,6 @@ class GroupSerializer(serializers.ModelSerializer):
     class Meta:
         model = Group
         fields = ['id', 'name']
-
-
-# FileUpload Parser
-class DeepSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Deep
-        fields = ('id', 'title', 'video_file')
